Checking code quality with Ruff (#2)
* cfg: Updating ruff used

* cfg: Some Ruff rules that work as it is

These codebase violates several best-practices on Python coding. I'm
adding a few recommended rules, but ignored or commented out the minimum
necessary to run as it is so we can handle it progressively in small
steps but secure as much rules as possible.

* cfg: GA checking code with Ruff

Don't check format for now. It would require too many changes that we
will postpone a little.

* cfg: Updating pixi.lock

* cfg: Updating pixi used in GA

* style:

* Updating pixi.lock

* cfg: pixi just frozen for now