Reconcile layer combination (#6)
* refact: Moving WeightsOfEvidence to root level

* refact: Reconciling 3D WeightsOfEvidence with common class

* test: Just import modules layer_combination (core, 2D, & 3D)

There was a bug in importing the dependencies that wasn't detected
before. This test covers that we won't miss it again in the future.

* fix: Typo on package name (transformation 2D)

* fix: Typo on package name (transformation 3D)

* fix: Relative path to import local transformation (2D & 3D)

* test: Also importing core `layer_combination` module

* test: get_w0

* feat: Unified get_w0()

* cfg: Adding hypothesis as requirement

* feat: Adding layer_combination.voter() copied from 2D

* test: test_validate_2D_voter()

Defining voter in unified layer_combination based on the 2D
implementation. This test validates that 2D implementation and unified
one returns the same answer in preparation to divert to a unified
.voter()

* test: test_voter_properties()

Validate behavior of unified voter()

* test: test_validate_3D_voter()

Confirm that legacy 3D voter gives the same result of the unified voter.

* refact, test: Reducing the possible testing sizes

This is just a concept check on multiple array sizes. It doesn't add a
huge array, but just make the calculation heavier.

* refact: Generalizing the 2D|3D operation

* refact: Diverting 2D & 3D to unified get_w0

The get_w0 is a trivial case and doesn't leave much room for doubts.

* test: Unified voter

* style:

* test: test_voter()

* style: