Reconciling `data_reader` (#11)
* refact: Initializing core data_readers with geopfa3d version

* refact: A transition GeospatialDataReaders class

While reconciling read_data into a single resource, let's keep this
entry point here during this transition period.

* style: Reformating with ruff

* clean: Removing identical 2D read_shapefile

* clean: Removing identical 2D read_shapefile

* clean: Removing identical 2D read_tif

* clean: Removing 2D read_csv, effectively identical

There where actually a couple of mistakes on the 2D version, but for
practical purposes, it is identical to the unified version.

* clean: Removing identical 2D gather_data

* clean: Removing identical 2D gather_processed_data

* clean: Removing identical 2D gather_exclusion_areas

* refact: Replacing 2D GeospatialDataReaders by unified code

A transition class with a warning to migrate to the reconciled 2D + 3D
code.

* clean: Unecessary encoding definition

* style: RET506, elif after raise

* style: PTH118, using pathlib.Path

* refact: Using starmap

* refact: replacing listdir by iterdir

* style: Moving fixed list to a set

* style:

* style: Ignoring a few rules

Let's ignore for now some rules, such as too many arguments for a
function, in specific places to postpone the need for larger refactoring
until we have a better suit of tests to protect the code.

* style:

* style: Breaking doc lines